+++ src/components/ingredient-details/detail.jsx
@@ -1,0 +1,12 @@
+import styles from './ingredient-details.module.css'
+
+const Detail = ({header, value}) => {
+  return (
+    <div className={`${styles.ingredientDetail} text_color_inactive`}>
+      <p className={`${styles.ingredientDetailText} text text_type_main-default`}>{header}</p>
+      <p className={`${styles.ingredientDetailText} text text_type_digits-default`}>{value}</p>
+    </div>
+  );
+}
+
+export default Detail;
