+++ src/components/ingredient-details/index.jsx
@@ -1,5 +1,6 @@
 import styles from './ingredient-details.module.css'
 import {useSelector} from "react-redux";
+import Detail from "./detail";
 
 const IngredientDetails = () => {
   const {
@@ -28,15 +29,6 @@ const IngredientDetails = () => {
         <Detail header='Углеводы, г' value={carbohydrates}/>
       </div>
     </>
-  );
-}
-
-const Detail = ({header, value}) => {
-  return (
-    <div className={`${styles.ingredientDetail} text_color_inactive`}>
-      <p className={`${styles.ingredientDetailText} text text_type_main-default`}>{header}</p>
-      <p className={`${styles.ingredientDetailText} text text_type_digits-default`}>{value}</p>
-    </div>
   );
 }
