--- src/components/modal-overlay/modal-overlay.module.css
@@ -1,8 +1,0 @@
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgb(0, 0, 0, 0.8);
-}
