+++ src/components/modal/modal.module.css
@@ -1,3 +1,12 @@
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgb(0, 0, 0, 0.8);
+}
+
 .modalContent {
   position: fixed;
   top: 50%;
